Add sharing session between processes per keyspace
It is enabled only for named grakn connections (because of easier way to find connection).

Behaviour before: open session, never closed. It is the quickest way of handling sessions,
and with benchmark I couldn't find the downsides of it (probably in long run, because of
memory leaks could be).

Behaviour after this change:

* For unnamed connections, it open a session and close it per transaction. It is twice slower,
than it was before.

* Named connections, share sessions per keyspace (opens only one session per connection). It is
1% slower, then behaiour before, but twice as fast, as open/close session per transaction.